Use SYSTEM and not REBOL
The historical equivalence between SYSTEM and REBOL is no longer.  The reason
is because it was deceptive when you would say `do [rebol [Type: 'Module] ...]`
or otherwise have header information there.  This could only be processed if a
LOAD was performed, e.g. if you used text like `do "rebol [Type 'Module] ..."`.
If the code was already loaded, what was actually happening was that your
object reference to `rebol` was evaluated and thrown away, then the BLOCK!
of your header information would be evaluated and thrown away.

Accordingly, Rebol has been made an error, informing you of this mistake...
that no LOAD-ed code should actually be executing a Rebol header signal.